Finished the Mecanum Teleop
I made all of my additions to the Mecanum class and created and completed the first part of our teleop.
One thing though is that we have to refactor our sensors, as the class structure was designed to emulate the different hardware assemblies, such as the drivetrain and the lift. However, by lumping all of the sensors into one class, it detaches from that nice feature.
On Monday we want to refactor the sensor class in about 30 minutes or so